Restructured docs components
