--- src/docs/_sass/_docs-block.scss
@@ -1,96 +1,0 @@
-.docs-block {
-  margin: 0 0 50px 0;
-  display: flex;
-
-  &__header {
-    color: $grayPrimary;
-    display: block;
-    font-size: 24px;
-    font-weight: $fontWeightNormal;
-    margin: 0;
-  }
-
-  &__info {
-    flex: 0 1 auto;
-    padding: 0 20px 0 0;
-    width: 230px;
-  }
-
-  &__content {
-    flex: 1 1 auto;
-
-    &--centered {
-      display: flex;
-      align-items: center
-    }
-
-    &--align-top {
-      & > * {
-        vertical-align: top;
-      }
-    }
-
-    &:hover {
-      position: relative;
-
-      .docs-highlight-placeholder {
-        outline: 2px dashed white;
-        box-shadow: black 0 0 10px 0;
-      }
-    }
-  }
-
-  &__contrast-box {
-    background-color: $grayPrimary;
-    padding: $baseline;
-    display: inline-block;
-  }
-
-  &__content-box {
-    padding: $baseline;
-    display: inline-block;
-  }
-}
-
-body.show-vr-lines {
-  .docs-block__content:hover {
-    &:before {
-      z-index: 1;
-      content: '';
-      background: linear-gradient(to bottom, #0ff 0, rgba(255, 255, 255, 0) 1px) repeat-y;
-      background-size: 100% rhythm(1);
-      height: 100%;
-      width: 100%;
-      position: absolute;
-      top: 0;
-      left: 0;
-      padding: 1px;
-      pointer-events: none;
-    }
-  }
-}
-
-body.show-holes {
-  .mint-bubble__hole,
-  .mint-box__hole,
-  .mint-promo-box__content,
-  .mint-header__left,
-  .mint-header__middle,
-  .mint-header__right,
-  .mint-overlayed-box,
-  .mint-overlayed-box__overlay,
-  .mint-footer__line,
-  .mint-content-box__header,
-  .mint-content-box__title,
-  .mint-content-box__footer,
-  .mint-content-box__content,
-  .mint-toplayer__heading,
-  .mint-toplayer__content,
-  .mint-toplayer__actions,
-  .mint-dropdown__hole,
-  .mint-dropdown__item-hole,
-  .mint-popup-menu__hole,
-  .mint-text-bit__hole {
-    @extend .docs-highlight-placeholder;
-  }
-}
