+++ src/docs/_sass/_navigation.scss
@@ -19,7 +19,7 @@
     width: 30px;
     height: 30px;
     margin-right: 5px;
-    background: center center url(../img/favicon-hd.png) no-repeat;
+    background: center center url(~/img/favicon-hd.png) no-repeat;
     background-size: contain;
   }
