--- src/docs/_sass/main.scss
@@ -1,104 +1,0 @@
-@charset "utf-8";
-
-@import "../../sass/config"; //style guide variables
-@import "../../sass/mixins";
-@import "../_sass/config";
-@import "../_sass/colors-list";
-@import "../_sass/color-box";
-@import "../_sass/docs-block";
-@import "../_sass/navigation";
-
-html, body {
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  padding: 50px 10px;
-}
-
-iframe {
-  border: none;
-  box-shadow: black 0 0 0 4px;
-}
-
-.data-container {
-  max-width: 960px;
-  margin: 0 auto;
-}
-
-.main-header {
-  margin: 20px 0 0;
-  font-size: 48px;
-  font-weight: $fontWeightBold;
-  color: $black;
-}
-
-.article-header {
-  color: $grayPrimary;
-  font-weight: $fontWeightNormal;
-  font-size: 36px;
-  padding: 0 0 60px;
-  margin: 15px 0;
-  border-bottom: 1px solid $grayPrimary;
-
-  &:hover {
-    .permalink {
-      display: inline-block;
-    }
-  }
-
-  .permalink {
-    display: none;
-    color: $grayPrimary;
-    text-decoration: none;
-    margin-left: 10px;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-}
-
-article {
-  margin-bottom: 100px;
-}
-
-.icons-list {
-  list-style: none;
-  padding: 0;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-start;
-
-  &__element {
-    width: 25%;
-    margin: 0 0 20px 0;
-    font-size: 16px;
-    line-height: 22px;
-    color: $black;
-    display: flex;
-    align-items: center;
-
-    &--wider {
-      width: 33%;
-    }
-
-    &--wide {
-      width: 50%;
-    }
-
-    &:before {
-      color: $mintPrimary;
-      font-size: 22px;
-      float: left;
-      margin: 0 10px 0 0;
-    }
-  }
-
-  &__element-info {
-    display: inline-block;
-    vertical-align: top;
-    margin: 21px 0 0;
-  }
-}
