+++ src/docs/components/docs-entry/docs-entry.scss
@@ -1,0 +1,30 @@
+@import '~sass/config';
+article {
+  margin-bottom: 100px;
+}
+
+.article-header {
+  color: $grayPrimary;
+  font-weight: $fontWeightNormal;
+  font-size: 36px;
+  padding: 0 0 60px;
+  margin: 15px 0;
+  border-bottom: 1px solid $grayPrimary;
+
+  &:hover {
+    .permalink {
+      display: inline-block;
+    }
+  }
+
+  .permalink {
+    display: none;
+    color: $grayPrimary;
+    text-decoration: none;
+    margin-left: 10px;
+
+    &:hover {
+      text-decoration: underline;
+    }
+  }
+}
