+++ src/entry.js
@@ -1,6 +1,5 @@
-import './docs/_sass/main.scss';
 import React from 'react';
 import { render } from 'react-dom';
-import Docs from './docs';
+import Docs from './docs/components/docs/index';
 
 render(<Docs/>, document.getElementById('root'));
